+++ pyvips/voperation.py
@@ -529,7 +529,7 @@ def cache_set_trace(trace):
 
 
 __all__ = [
-    'Operation', 
-    'cache_set_max', 'cache_set_max_mem', 'cache_set_max_files', 
+    'Operation',
+    'cache_set_max', 'cache_set_max_mem', 'cache_set_max_files',
     'cache_set_trace'
 ]