+++ CHANGELOG.rst
@@ -6,6 +6,7 @@
 * add `version()` to get libvips version number
 * add `cache_set_max()`, `cache_set_max_mem()`, `cache_set_max_files()`, 
   `cache_set_trace()`
+* all glib log levels sent to py logger
 
 ## Version 2.0.1 (23 August 2017)